% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-Characteristics.R
\name{Characteristics}
\alias{Characteristics}
\title{Characterisitcs Module Class}
\description{
Characteristics module that shows characteristics results (table and plot) from the CohortCharacteristics package.
}
\examples{
{
\donttest{
 library(DarwinShinyModules)

 if (
   require(
     "CohortCharacteristics",
     character.only = TRUE,
     quietly = TRUE,
     warn.conflicts = FALSE
   )
 ) {
    result <- omopgenerics::importSummarisedResult(system.file(
      package = "DarwinShinyModules",
      "dummyData/CohortCharacteristics/1.1.1/characteristics.csv"
    ))

    charMod <- Characteristics$new(result = result)

    ui <- shiny::fluidPage(
      charMod$UI()
    )

    server <- function(input, output, session) {
      charMod$server(input, output, session)
    }

    if (interactive()) {
      shiny::shinyApp(ui = ui, server = server)
    }
  }
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{Characteristics}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCharacteristics()}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-Characteristics-initialize}{\code{Characteristics$new()}}
    \item \href{#method-Characteristics-clone}{\code{Characteristics$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristics-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristics-initialize}{}}}
\subsection{\code{Characteristics$new()}}{
  Initializer method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Characteristics$new(result = NULL, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCharacteristics()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Characteristics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Characteristics-clone}{}}}
\subsection{\code{Characteristics$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{Characteristics$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
