% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-SummarizedResult.R
\name{checkCDMNames}
\alias{checkCDMNames}
\title{checkCDMNames}
\usage{
checkCDMNames(result, .softValidation)
}
\arguments{
\item{result}{(\code{result}) Result to check CDM names for.}

\item{.softValidation}{(\code{logical(1)}) If the check fails, throws a warning, rather than an error.}
}
\value{
When the check passes returns \code{invisible(TRUE)}, otherwise returns the error
}
\description{
Checks the CDM names in a result. You can checkout the valid options running: \code{getCDMAcronyms()}.
}
\examples{
if (interactive()) {
  checkCDMNames(result)
}
}
