% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortAttrition.R
\name{CohortAttrition}
\alias{CohortAttrition}
\title{CohortAttrition Module Class}
\value{
\code{self}
}
\description{
CohortAttrition module that displays tables and plots of the
\code{summarised_result} object created by \code{CohortCharacteristics::summariseCohortAttrition()}.
}
\examples{
if (interactive()) {
  CDMConnector::requireEunomia()
  con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomiaDir())
  cdm <- CDMConnector::cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")

  cdm <- DrugUtilisation::generateIngredientCohortSet(cdm = cdm, name = "my_cohort", ingredient = c("warfarin", "acetaminophen"))

  summarisedAttrition <- CohortCharacteristics::summariseCohortAttrition(cdm$my_cohort)

  mod <- Attrition$new(result = summarisedAttrition)
  DarwinShinyModules::preview(mod)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{CohortAttrition}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{(\code{summarised_result})}

\item{\code{cdmNames}}{(\code{character(n)})}

\item{\code{cohortNames}}{(\code{character(n)})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortAttrition-new}{\code{CohortAttrition$new()}}
\item \href{#method-CohortAttrition-clone}{\code{CohortAttrition$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>DarwinShinyModules::ShinyModule$getReactiveValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortAttrition-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortAttrition-new}{}}}
\subsection{Method \code{new()}}{
Initializer method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortAttrition$new(result, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{(\code{summarised_result}) Object created by \code{CohortCharacteristics::summariseLargeScaleCharacteristics()}.}

\item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortAttrition-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortAttrition-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortAttrition$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
