% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortMenuItem.R
\name{CohortMenuItem}
\alias{CohortMenuItem}
\title{Cohort menu item}
\description{
CohortMenuItem Module
}
\examples{
library(DarwinShinyModules)

cohortMenuItem <- CohortMenuItem$new(appId = "id", data = mtcars)

if (interactive()) {
  preview(cohortMenuItem)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{CohortMenuItem}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortTable}}{(\code{Table}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortMenuItem-new}{\code{CohortMenuItem$new()}}
\item \href{#method-CohortMenuItem-validate}{\code{CohortMenuItem$validate()}}
\item \href{#method-CohortMenuItem-UI}{\code{CohortMenuItem$UI()}}
\item \href{#method-CohortMenuItem-tabItem}{\code{CohortMenuItem$tabItem()}}
\item \href{#method-CohortMenuItem-server}{\code{CohortMenuItem$server()}}
\item \href{#method-CohortMenuItem-clone}{\code{CohortMenuItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to display in the table, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-UI"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$UI(text = "Cohorts", subItemText = "Cohort attrition")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{(\code{character(1)}) Menu item text.}

\item{\code{subItemText}}{(\code{character(1)}) Sub menu item text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag}
tabItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-tabItem"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-tabItem}{}}}
\subsection{Method \code{tabItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$tabItem(tableTitle = "Cohort attrition")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tableTitle}}{(\code{character(1)}) Title to use for the cohort table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{tabItem}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-server"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortMenuItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortMenuItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortMenuItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
