% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortOverlap.R
\name{CohortOverlap}
\alias{CohortOverlap}
\title{CohortOverlap Module Class}
\description{
CohortOverlap module that shows cohort overlap results (table and plot) from the CohortCharacteristics package.
}
\examples{
{
if (interactive()) {
  library(CohortCharacteristics)
  library(DarwinShinyModules)

  cdm <- mockCohortCharacteristics()

  result <- summariseCohortOverlap(cdm$cohort2)

  mod <- CohortOverlap$new(result)

  preview(mod)
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{CohortOverlap}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCohortOverlap()}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CohortOverlap-initialize}{\code{CohortOverlap$new()}}
    \item \href{#method-CohortOverlap-clone}{\code{CohortOverlap$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortOverlap-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CohortOverlap-initialize}{}}}
\subsection{\code{CohortOverlap$new()}}{
  Initializer method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CohortOverlap$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCohortOverlap()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortOverlap-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortOverlap-clone}{}}}
\subsection{\code{CohortOverlap$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CohortOverlap$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
