% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortTab.R
\name{CohortTab}
\alias{CohortTab}
\title{Cohort tab}
\description{
CohortTab Module
}
\examples{
library(DarwinShinyModules)

cohortTab <- CohortTab$new(appId = "id", data = mtcars)

if (interactive()) {
  preview(cohortTab)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{CohortTab}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{cohortTable}}{(\code{Table}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortTab-new}{\code{CohortTab$new()}}
\item \href{#method-CohortTab-validate}{\code{CohortTab$validate()}}
\item \href{#method-CohortTab-UI}{\code{CohortTab$UI()}}
\item \href{#method-CohortTab-server}{\code{CohortTab$server()}}
\item \href{#method-CohortTab-clone}{\code{CohortTab$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTab-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTab-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTab$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTab-validate"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTab-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTab$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTab-UI"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTab-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTab$UI(tabName = "cohort_attrition", tableTitle = "Cohort attrition")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tabName}}{(\code{character(1)}) Title to use for the tab.}

\item{\code{tableTitle}}{(\code{character(1)}) Title to use for the cohort table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag.list}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTab-server"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTab-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTab$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTab-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTab-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortTab$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
