% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortTiming.R
\name{CohortTiming}
\alias{CohortTiming}
\title{CohortTiming Module Class}
\description{
CohortTiming module that shows cohort timing results (table and plot) from the CohortCharacteristics package.
}
\examples{
{
if (interactive()) {
  library(CohortCharacteristics)
  library(omock)
  library(DrugUtilisation)

  cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

  cdm <- generateIngredientCohortSet(
    cdm = cdm,
    name = "my_cohort",
    ingredient = c("acetaminophen", "morphine", "warfarin")
  )

  result <- summariseCohortTiming(cdm$my_cohort)

  mod <- CohortTiming$new(result)

  preview(mod)
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{CohortTiming}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCohortTiming()}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CohortTiming-initialize}{\code{CohortTiming$new()}}
    \item \href{#method-CohortTiming-clone}{\code{CohortTiming$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTiming-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTiming-initialize}{}}}
\subsection{\code{CohortTiming$new()}}{
  Initializer method
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CohortTiming$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Result of \code{CohortCharacteristics::summariseCohortTiming()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortTiming-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortTiming-clone}{}}}
\subsection{\code{CohortTiming$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CohortTiming$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
