% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateApp.R
\name{createApp}
\alias{createApp}
\title{Function that creates a stand-alone shiny app from a given app structure.}
\usage{
createApp(
  path,
  appStructure,
  theme,
  additionalFiles = c(),
  createProject = FALSE,
  openProject = FALSE
)
}
\arguments{
\item{path}{path where shiny app will be stored}

\item{appStructure}{application structure as a list}

\item{theme}{theme as a character}

\item{additionalFiles}{optional vector of files to be copied}

\item{createProject}{if the path should be made into a project}

\item{openProject}{if, after creation of the project, the project should be opened}
}
\description{
Function that creates a stand-alone shiny app from a given app structure.
}
\examples{
library(DarwinShinyModules)

base <- Text$new("**base**")
nested_a <- Text$new("**nested A**")
nested_b <- Text$new("**nested B**")
sub_a <- Text$new("**sub A**")
sub_b <- Text$new("**sub B**")
comb_a <- Text$new("**comb A**")
comb_b <- Text$new("**comb B**")
comb_c <- Text$new("**comb C**")

appStructure <- list(
   base = base,
   nested = list(nested_a, nested_b),
   nested_sub = list(
     sub_a = sub_a,
     sub_b = sub_b
   ),
   nested_combined = list(
     comb_a_b = list(comb_a, comb_b),
     comb_c = comb_c
   )
 )

if (interactive()) {
  createApp(path = tempdir(),
            appStructure = appStructure,
            theme = 'shinymodules-darwin')
}
}
