% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darwin-Utils.R
\name{darwinFooter}
\alias{darwinFooter}
\title{darwinFooter}
\usage{
darwinFooter(type = "shinydashboard", deployDate = Sys.Date())
}
\arguments{
\item{type}{for which app type the footer should be generated}

\item{deployDate}{Date when the app is deployed}
}
\value{
\code{tagList}
}
\description{
Creates the DARWIN EU (c) footer.
}
\examples{
ui <- shiny::fluidPage(
  darwinHeader(),
  shiny::p("Content"),
  darwinFooter()
)

server <- function(input, output, session) {}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
