% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appBuilders.R
\name{dashboardApp}
\alias{dashboardApp}
\title{dashboardApp}
\usage{
dashboardApp(moduleList, title = NULL)
}
\arguments{
\item{moduleList}{(\code{list(list())}) A list of named lists, containing modules.
The level of nesting groups or separates modules in menu items \code{"_"} will be read as a space.}

\item{title}{(\code{character(1)}: \code{NULL}) Title of the app}
}
\value{
\code{NULL}
}
\description{
dashboardApp
}
\examples{
library(DarwinShinyModules)
library(ggplot2)

mtcarsTable <- Table$new(data = mtcars)
irisTable <- Table$new(data = iris)

irisPlotFun <- function(data) {
  ggplot(data = data, mapping = aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
    geom_point() +
    theme_bw()
}

irisPlot <- PlotStatic$new(fun = irisPlotFun, args = list(data = iris))

if (interactive()) {
  dashboardApp(
    moduleList = list(
      # Menu item `Iris`
      Iris = list(irisPlot, irisTable),
      # Menu item `MT Cars`
      MT_Cars = list(mtcarsTable)
    )
  )
}
}
