% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseMenuItem.R
\name{DatabaseMenuItem}
\alias{DatabaseMenuItem}
\title{Database menu item}
\description{
DatabaseMenuItem Module
}
\examples{
library(DarwinShinyModules)

DatabaseMenuItem <- DatabaseMenuItem$new(appId = "id", data = mtcars)

if (interactive()) {
  preview(DatabaseMenuItem)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{DatabaseMenuItem}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dbTable}}{(\code{Table}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseMenuItem-new}{\code{DatabaseMenuItem$new()}}
\item \href{#method-DatabaseMenuItem-validate}{\code{DatabaseMenuItem$validate()}}
\item \href{#method-DatabaseMenuItem-UI}{\code{DatabaseMenuItem$UI()}}
\item \href{#method-DatabaseMenuItem-tabItem}{\code{DatabaseMenuItem$tabItem()}}
\item \href{#method-DatabaseMenuItem-server}{\code{DatabaseMenuItem$server()}}
\item \href{#method-DatabaseMenuItem-clone}{\code{DatabaseMenuItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to display in the table, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-UI"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$UI(text = "Databases", subItemText = "Database details")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{(\code{character(1)}) Menu item text.}

\item{\code{subItemText}}{(\code{character(1)}) Sub menu item text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag}
tabItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-tabItem"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-tabItem}{}}}
\subsection{Method \code{tabItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$tabItem(tableTitle = "Study databases")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tableTitle}}{(\code{character(1)}) Title to use for the cohort table.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{tabItem}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-server"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseMenuItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseMenuItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseMenuItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
