% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseOverview.R
\name{DatabaseOverview}
\alias{DatabaseOverview}
\title{DatabaseOverview Module Class}
\description{
The DatabaseOverview module displays information about data partners, and
links to the Portal and data partner website.
}
\examples{
library(DarwinShinyModules)

databaseOverview <- DatabaseOverview$new("IPCI", "CPRD GOLD", "SIDIAP")

if (interactive()) {
  preview(databaseOverview)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{DatabaseOverview}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{table}}{(\code{Table}) Table module}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseOverview-new}{\code{DatabaseOverview$new()}}
\item \href{#method-DatabaseOverview-clone}{\code{DatabaseOverview$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseOverview-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseOverview-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseOverview$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{character(n)}) Individual database names. See Details for
the supported data partners.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
\describe{
\item{A}{"APHM"}
\item{B}{"BIFAP"}
\item{C}{"CDW Bordeaux"\cr"CPRD GOLD" "CRN"}
\item{D}{"DK-DHR"}
\item{E}{"EBB"\cr"EMDB - ULSEDV"\cr"EMDB - ULSGE"\cr"EMDB - ULSRA"}
\item{F}{"FinOMOP - ACI Varha"\cr"FinOMOP - HILMO"\cr"FinOMOP - HUS"\cr"FinOMOP - TaUH Pirha"}
\item{H}{"H12O"\cr"HARMONY Platform"\cr"HARMONY - ALL"\cr"HARMONY - AML"\cr"HARMONY - CML"\cr"HARMONY - MM"}
\item{I}{"IMASIS"\cr"InGef RDB"\cr"IPCI"\cr"IQVIA DA Germany"\cr"IQVIA LPD Belgium"}
\item{N}{"NAJS"\cr"NCR"\cr"NLHR"\cr"NLHR@UiO:PERINATAL"\cr"NNRD"}
\item{P}{"PGH"\cr"PRISIB"\cr"POLIMI"}
\item{S}{"SIDIAP"\cr"SNDS"\cr"SUCD"}
\item{U}{"UKBB"\cr"ULSM-RT"}
\item{V}{"VID"}
}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseOverview-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseOverview-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseOverview$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
