% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deploy.R
\name{deployAppStructure}
\alias{deployAppStructure}
\title{deployAppStructure}
\usage{
deployAppStructure(appStructure, appDir = tempfile(), ...)
}
\arguments{
\item{appStructure}{(\code{list}: \code{NULL}) appStructure object, i.e. a named list
of ShinyModule objects or (named) lists containing (named) ShinyModule
objects. Representing the navigation (side) bar of the shiny app.}

\item{appDir}{(\code{character}) File path to the directory where app files will
be written to, and the app is being deployed from. See \code{rsconnect::deployApp}
for further details.}

\item{...}{Arguments for \code{rsconnect::deployApp}}
}
\value{
\code{NULL}
}
\description{
Wrapper for \code{rsconnect::deployApp}. Deploys a shiny app from the appStructure
}
