% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DoseCoverage.R
\name{DoseCoverage}
\alias{DoseCoverage}
\title{DoseCoverage Module Class}
\description{
DoseCoverage module that shows a table
}
\examples{
{
if (interactive()) {
  library(DarwinShinyModules)
  library(DrugUtilisation)

  cdm <- DrugUtilisation::mockDrugUtilisation(numberIndividuals = 100)

  res <- cdm$cohort1 |>
    DrugUtilisation::summariseProportionOfPatientsCovered(followUpDays = 365)

  mod <- DoseCoverage$new(res)
  preview(mod)
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{DoseCoverage}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result}) Result}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DoseCoverage-initialize}{\code{DoseCoverage$new()}}
    \item \href{#method-DoseCoverage-clone}{\code{DoseCoverage$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoseCoverage-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DoseCoverage-initialize}{}}}
\subsection{\code{DoseCoverage$new()}}{
  Initializer method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DoseCoverage$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Object created by \code{DrugUtilisation::summariseDoseCoverage()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DoseCoverage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DoseCoverage-clone}{}}}
\subsection{\code{DoseCoverage$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DoseCoverage$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
