% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugExposureDiagnostics.R
\name{DrugExposureDiagnostics}
\alias{DrugExposureDiagnostics}
\title{DrugExposureDiagnostics Module Class}
\description{
DrugExposureDiagnostics module that shows tables and plots
}
\details{
The module consists of the following:
\describe{
\item{"dataPlotPanel"}{Table and a plot (bar or box) for each check.}
\item{"metaDataPanel"}{Table containing the metadata.}
}
}
\examples{
{
library(DarwinShinyModules)

if (require("DrugExposureDiagnostics", character.only = TRUE, quietly = TRUE,
warn.conflicts = FALSE)) {
  ded <- readRDS(system.file(package = "DarwinShinyModules",
                             "dummyData/DrugExposureDiagnostics/1.1.1/ded.rds"))
  mod <- DrugExposureDiagnostics$new(resultList = ded)
  if (interactive()) {
    preview(mod)
  }
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{DrugExposureDiagnostics}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrugExposureDiagnostics-new}{\code{DrugExposureDiagnostics$new()}}
\item \href{#method-DrugExposureDiagnostics-clone}{\code{DrugExposureDiagnostics$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrugExposureDiagnostics-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrugExposureDiagnostics-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrugExposureDiagnostics$new(resultList, database_id = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultList}}{(\code{list}) List containing the output of the checks}

\item{\code{database_id}}{(\code{character}) Database identifier (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrugExposureDiagnostics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrugExposureDiagnostics-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrugExposureDiagnostics$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
