% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-DT.R
\name{DTTable}
\alias{DTTable}
\title{DTTable Module Class}
\description{
GTTable module that displays tables using \code{DT} that are supported by
\code{DT::renderDT()} and \code{DT::DTOutput()}.
}
\examples{
library(DarwinShinyModules)

dtTable <- DTTable$new(
  fun = DT::datatable,
  args = list(data = iris)
)

if (interactive()) {
  preview(dtTable)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{DTTable}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{fun}}{(\code{function}) Function to produce a \code{gt} table with, i.e \code{gt::gt}.}

    \item{\code{args}}{(\code{list}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DTTable-initialize}{\code{DTTable$new()}}
    \item \href{#method-DTTable-clone}{\code{DTTable$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DTTable-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DTTable-initialize}{}}}
\subsection{\code{DTTable$new()}}{
  Initializer method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DTTable$new(fun, args, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{fun}}{(\code{function}) Function to produce a \code{gt} table with, i.e \code{gt::gt}.}
      \item{\code{args}}{(\code{list()}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DTTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DTTable-clone}{}}}
\subsection{\code{DTTable$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DTTable$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
