% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-EventDuration.R
\name{EventDuration}
\alias{EventDuration}
\title{EventDuration}
\description{
Internal module that displays the Event Duration from the
\code{TreatmentPatterns} package.
}
\examples{
{
  if (interactive()) {
    library(DarwinShinyModules)

    tpr <- TreatmentPatterns::TreatmentPatternsResults$new(
      filePath = "./inst/dummyData/TreatmentPatterns/3.0.0"
    )

    treatmentPathways <- EventDuration$new(
      treatmentPathways = tpr$treatment_pathways,
      dmSourceInfo = tpr$cdm_source_info
    )

    preview(treatmentPathways)
  }
}

}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{EventDuration}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-EventDuration-new}{\code{EventDuration$new()}}
\item \href{#method-EventDuration-clone}{\code{EventDuration$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventDuration-new"></a>}}
\if{latex}{\out{\hypertarget{method-EventDuration-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventDuration$new(summaryEventDuration, cdmSourceInfo, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-EventDuration-clone"></a>}}
\if{latex}{\out{\hypertarget{method-EventDuration-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{EventDuration$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
