% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table-Flextable.R
\name{Flextable}
\alias{Flextable}
\title{Flextable Module Class}
\value{
\code{self}
}
\description{
Flextable module that displays tables using \code{flextable} that are displayed by
\code{renderUI} and \code{uiOutput}.
}
\examples{
library(DarwinShinyModules)

gtTable <- Flextable$new(
  fun = flextable::flextable,
  args = list(data = iris)
)

if (interactive()) {
  preview(gtTable)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{Flextable}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fun}}{(\code{function}) Function to produce a \code{flextable} table with, i.e \code{flextable::flextable}.}

\item{\code{args}}{(\code{list}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Flextable-new}{\code{Flextable$new()}}
\item \href{#method-Flextable-clone}{\code{Flextable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>DarwinShinyModules::ShinyModule$getReactiveValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flextable-new"></a>}}
\if{latex}{\out{\hypertarget{method-Flextable-new}{}}}
\subsection{Method \code{new()}}{
Initializer method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flextable$new(fun, args, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{(\code{function}) Function to produce a \code{flextable} table with, i.e \code{flextable::flextable}.}

\item{\code{args}}{(\code{list()}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}

\item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Flextable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Flextable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Flextable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
