% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCDMAcronyms}
\alias{getCDMAcronyms}
\title{getCDMAcronyms}
\usage{
getCDMAcronyms()
}
\value{
\code{character(n)}
}
\description{
Returns a character vector containing all the data source acronyms from the DARWIN EU Portal
}
\examples{
getCDMAcronyms()
}
