% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTTable.R
\name{GTTable}
\alias{GTTable}
\title{GTTable Module Class}
\value{
\code{self}
}
\description{
GTTable module that displays tables using \code{gt} that are supported by
\code{gt::render_gt()} and \code{gt::gt_output()}.
}
\examples{
library(DarwinShinyModules)

gtTable <- GTTable$new(
  fun = gt::gt,
  args = list(data = iris)
)

if(interactive()) {
  preview(gtTable)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{GTTable}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{fun}}{(\code{function}) Function to produce a \code{gt} table with, i.e \code{gt::gt}.}

\item{\code{args}}{(\code{list}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-GTTable-new}{\code{GTTable$new()}}
\item \href{#method-GTTable-clone}{\code{GTTable$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GTTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-GTTable-new}{}}}
\subsection{Method \code{new()}}{
Initializer method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GTTable$new(fun, args)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{(\code{function}) Function to produce a \code{gt} table with, i.e \code{gt::gt}.}

\item{\code{args}}{(\code{list()}) Arguments for said function as a named list i.e. \code{list(data = iris)}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-GTTable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-GTTable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GTTable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
