% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Incidence.R
\name{Incidence}
\alias{Incidence}
\title{Incidence}
\description{
Incidence Module
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{Incidence}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Incidence-new}{\code{Incidence$new()}}
\item \href{#method-Incidence-UI}{\code{Incidence$UI()}}
\item \href{#method-Incidence-server}{\code{Incidence$server()}}
\item \href{#method-Incidence-clone}{\code{Incidence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Incidence-new"></a>}}
\if{latex}{\out{\hypertarget{method-Incidence-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Incidence$new(appId, incidence)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Incidence-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Incidence-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Incidence$UI(title = "Table")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Incidence-server"></a>}}
\if{latex}{\out{\hypertarget{method-Incidence-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Incidence$server(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Incidence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Incidence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Incidence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
