% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidenceMenuItem.R
\name{IncidenceMenuItem}
\alias{IncidenceMenuItem}
\title{Incidence menu item}
\description{
IncidenceMenuItem Module
}
\examples{
library(DarwinShinyModules)

incidenceMenuItem <- IncidenceMenuItem$new(appId = "id", data = mtcars)

if (interactive()) {
  preview(incidenceMenuItem)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{IncidenceMenuItem}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IncidenceMenuItem-new}{\code{IncidenceMenuItem$new()}}
\item \href{#method-IncidenceMenuItem-validate}{\code{IncidenceMenuItem$validate()}}
\item \href{#method-IncidenceMenuItem-UI}{\code{IncidenceMenuItem$UI()}}
\item \href{#method-IncidenceMenuItem-tabItem}{\code{IncidenceMenuItem$tabItem()}}
\item \href{#method-IncidenceMenuItem-server}{\code{IncidenceMenuItem$server()}}
\item \href{#method-IncidenceMenuItem-clone}{\code{IncidenceMenuItem$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-new"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to display, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-validate"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-UI"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$UI(
  text = "Incidence",
  tabName = "incidence",
  subItemText = "incidence estimates",
  subItemTabName = "incidence_estimates"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{text}}{(\code{character(1)}) Menu item text.}

\item{\code{tabName}}{(\code{character(1)}) Title to use for the tab.}

\item{\code{subItemText}}{(\code{character(1)}) Sub menu item text.}

\item{\code{subItemTabName}}{(\code{character(1)}) Sub menu item text.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag}
tabItem
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-tabItem"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-tabItem}{}}}
\subsection{Method \code{tabItem()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$tabItem(
  tabName = "incidence_estimates",
  title = "Incidence estimates - general population",
  addTreatmentPicker = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tabName}}{(\code{character(1)}) Title to use for the tab.}

\item{\code{title}}{(\code{character(1)}) Title to use as heading.}

\item{\code{addTreatmentPicker}}{if a picker needs to be added containing treatment}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{tabItem}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-server"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidenceMenuItem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IncidenceMenuItem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidenceMenuItem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
