% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractivePlots.R
\name{InteractivePlots}
\alias{InteractivePlots}
\title{InteractivePlots}
\description{
Class to handle the interactive plots of TreatmentPatterns (Sunburst plot &
Sankey diagram)
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{InteractivePlots}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InteractivePlots-UI}{\code{InteractivePlots$UI()}}
\item \href{#method-InteractivePlots-server}{\code{InteractivePlots$server()}}
\item \href{#method-InteractivePlots-clone}{\code{InteractivePlots$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="initialize"><a href='../../DarwinShinyModules/html/Module.html#method-Module-initialize'><code>DarwinShinyModules::Module$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlots-UI"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlots-UI}{}}}
\subsection{Method \code{UI()}}{
Method to include a \link[shinydashboard]{tabItem} to include the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlots$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{tabItem})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlots-server"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlots-server}{}}}
\subsection{Method \code{server()}}{
Method to handle the back-end.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlots$server(input, output, session, inputHandler)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})\cr
Input from the server function.}

\item{\code{output}}{(\code{output})\cr
Output from the server function.}

\item{\code{session}}{(\code{session})\cr
Session from the server function.}

\item{\code{inputHandler}}{(\code{inputHandler})\cr
\link[TreatmentPatterns]{InputHandler} class.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InteractivePlots-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InteractivePlots-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InteractivePlots$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
