% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CohortCharacteristics-LargeScaleCharacteristics.R
\name{LargeScaleCharacteristics}
\alias{LargeScaleCharacteristics}
\title{LargeScaleCharacteristics Module Class}
\description{
LargeScaleCharacteristics module that displays tables and plots of the
\code{summarised_result} object created by \code{CohortCharacteristics::summariseLargeScaleCharacteristics()}.
}
\examples{
if (interactive()) {
  CDMConnector::requireEunomia()
   con <- DBI::dbConnect(duckdb::duckdb(), CDMConnector::eunomiaDir())
   cdm <- CDMConnector::cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")

   cdm <- DrugUtilisation::generateIngredientCohortSet(
     cdm = cdm,
     name = "my_cohort",
     ingredient = c("warfarin", "acetaminophen")
  )

   cdm$my_cohort <- cdm$my_cohort |>
     PatientProfiles::addAge(ageGroup = list(
      `0 to 17` = c(0, 17),
      `>=18` = c(18, Inf)
    )) |>
    PatientProfiles::addSex()

  result <- CohortCharacteristics::summariseLargeScaleCharacteristics(
    cohort = cdm$my_cohort,
    eventInWindow = "condition_occurrence",
    strata = list("age_group", "sex")
  )

  lscMod <- LargeScaleCharacteristics$new(result = result)
  DarwinShinyModules::preview(lscMod)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{LargeScaleCharacteristics}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result})}

    \item{\code{cdmNames}}{(\code{character(n)})}

    \item{\code{cohortNames}}{(\code{character(n)})}

    \item{\code{strata}}{(\code{character(n)})}

    \item{\code{windows}}{(\code{character(n)})}

    \item{\code{table}}{(\code{ShinyModule})}

    \item{\code{tableTop}}{(\code{ShinyModule})}

    \item{\code{plot}}{(\code{ShinyModule})}

    \item{\code{plotCompared}}{(\code{ShinyModule})}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LargeScaleCharacteristics-initialize}{\code{LargeScaleCharacteristics$new()}}
    \item \href{#method-LargeScaleCharacteristics-clone}{\code{LargeScaleCharacteristics$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LargeScaleCharacteristics-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LargeScaleCharacteristics-initialize}{}}}
\subsection{\code{LargeScaleCharacteristics$new()}}{
  Initializer method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LargeScaleCharacteristics$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Object created by \code{CohortCharacteristics::summariseLargeScaleCharacteristics()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LargeScaleCharacteristics-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LargeScaleCharacteristics-clone}{}}}
\subsection{\code{LargeScaleCharacteristics$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LargeScaleCharacteristics$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
