% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchFromDisk.R
\name{launchFromDisk}
\alias{launchFromDisk}
\title{launchFromDisk}
\usage{
launchFromDisk(
  filePath,
  launchFun = DarwinShinyModules::launchDarwinBslibApp,
  daemons = 2
)
}
\arguments{
\item{filePath}{(\code{character}: \code{"./appStructure.qs"}) Path to the qs-file,
containing the \code{appStructure}, or where to write the \code{appStructure} to.}

\item{launchFun}{(\code{launchFun}: \code{DarwinShinyModules::launchDarwinBslibApp})
launch function to use, may be any function that launches a shiny app from
an appStructure object.}

\item{daemons}{(\code{numeric}: 2) Number of sub R proccess to kick off to load
in results asynchronously.}
}
\value{
\code{shiny.appobj} when the app is found, otherwise returns \code{NULL} invisible
}
\description{
Launches the shiny app from appStructure object from either a qs-file, or
from the object directly, using the provided \code{launchFun}. When the qs-file
does not exist, but the appStructure does, the appStructure is saved to
disk, to the provided \code{fileName}.
}
