% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAppStructure.R
\name{loadAppStructure}
\alias{loadAppStructure}
\title{loadAppStructure}
\usage{
loadAppStructure(
  filePath,
  appStructureFileName = "appStructure.qs",
  subDir = NULL
)
}
\arguments{
\item{filePath}{(\code{character(1)}) Path to the tar.gz file downlaoded from
posit connect.}

\item{appStructureFileName}{(\code{character(1)}) Name of the appStructure file.}

\item{subDir}{(\code{character(1)}: \code{NULL}) Optional deviation in the sub
directory, if the tar.gz is not structured identically to how posit connect
structures these files.}
}
\value{
\code{list} appStructure
}
\description{
Loads an \code{appStructure} object from a downloaded tar.gz file.
}
