% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSourceTable.R
\name{makeSourceTable}
\alias{makeSourceTable}
\title{makeSourceTable}
\usage{
makeSourceTable(
  cdm_source_name = NA,
  cdm_source_abbreviation = NA,
  cdm_holder = NA,
  source_description = NA,
  source_documentation_reference = NA,
  cdm_etl_reference = NA,
  source_release_date = as.Date(NA),
  cdm_release_date = as.Date(NA),
  cdm_version = NA,
  cdm_version_concept_id = NA,
  vocabulary_version = NA
)
}
\arguments{
\item{cdm_source_name}{(\code{character(n)}: \code{NA}) Required}

\item{cdm_source_abbreviation}{(\code{character(n)}: \code{NA}) Required}

\item{cdm_holder}{(\code{character(n)}: \code{NA}) Required}

\item{source_description}{(\code{character(n)}: \code{NA}) Optional}

\item{source_documentation_reference}{(\code{character(n)}: \code{NA}) Optional}

\item{cdm_etl_reference}{(\code{character(n)}: \code{NA}) Optional}

\item{source_release_date}{(\code{Date(n)}: \code{NA}) Required}

\item{cdm_release_date}{(\code{Date(n)}: \code{NA}) Required}

\item{cdm_version}{(\code{character(n)}: \code{NA}) Optional}

\item{cdm_version_concept_id}{(\code{numeric(n)}: \code{NA}) Required (As of CDM 5.4)}

\item{vocabulary_version}{(\code{character(n)}: \code{NA}) Required}
}
\description{
Function to make a cdm_source table. Found here:
https://ohdsi.github.io/CommonDataModel/cdm54.html#cdm_source
}
\examples{
makeSourceTable(
  cdm_source_name = "Dummy CMD",
  cdm_source_abbreviation = "Dummy",
  cdm_holder = "Dummy organisation",
  source_description = "A dummy table",
  source_documentation_reference = "",
  cdm_etl_reference = "",
  source_release_date = as.Date("1970-01-01"),
  cdm_release_date = as.Date("1970-01-01"),
  cdm_version = "",
  cdm_version_concept_id = "",
  vocabulary_version = ""
)
}
