% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns.R
\name{makeTreatmentPatternsModule}
\alias{makeTreatmentPatternsModule}
\title{makeTreatmentPatternsModule}
\usage{
makeTreatmentPatternsModule(treatmentPathways, version = NULL)
}
\arguments{
\item{treatmentPathways}{(\code{data.frame}) treatmentPathways result.}

\item{version}{(\code{character(1)}: \code{NULL}) Main version to use i.e. \code{"2"},
\code{"3"}, or \code{NULL} which determines the installed version automatically.}
}
\value{
\code{TreatmentPatternsResults}
}
\description{
Factory function to create a TreatmentPatterns module. Supports 2.x.x and
3.x.x versions of TreatmentPatterns
}
\examples{
d1 <- read.csv(system.file(package = "DarwinShinyModules", "dummyData/TreatmentPatterns/2.7.0/treatmentPathways.csv")
d2 <- read.csv(system.file(package = "DarwinShinyModules", "dummyData/TreatmentPatterns/3.0.0/treatment_pathways.csv")

mod1 <- makeTreatmentPatternsModule(d1, version = "2")
mod2 <- makeTreatmentPatternsModule(d2, version = "3")

if (interactive()) {
  preview(list(mod1, mod2))
}
}
