% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-Characteristics.R
\name{moduleCharacteristics}
\alias{moduleCharacteristics}
\title{moduleCharacteristics}
\usage{
moduleCharacteristics(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCharacteristics()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCharacteristics
}
\examples{
if (interactive()) {
  moduleCharacteristics(result)
}
}
