% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortAttrition.R
\name{moduleCohortAttrition}
\alias{moduleCohortAttrition}
\title{moduleCohortAttrition}
\usage{
moduleCohortAttrition(result, .softValidation)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortAttrition} function from the CohortCharacteristics pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCohortAttrition
}
\examples{
if (interactive()) {
  moduleCohortAttrition(result)
}
}
