% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortCodelist.R
\name{moduleCohortCodelist}
\alias{moduleCohortCodelist}
\title{moduleCohortCodelist}
\usage{
moduleCohortCodelist(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortCodelist()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCohortCodelist
}
\examples{
if (interactive()) {
  moduleCohortCodelist(result)
}
}
