% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortCount.R
\name{moduleCohortCount}
\alias{moduleCohortCount}
\title{moduleCohortCount}
\usage{
moduleCohortCount(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortCount()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCohortCount
}
\examples{
if (interactive()) {
  moduleCohortCount(result)
}
}
