% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortOverlap.R
\name{moduleCohortOverlap}
\alias{moduleCohortOverlap}
\title{moduleCohortOverlap}
\usage{
moduleCohortOverlap(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortOverlap()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCohortOverlap
}
\examples{
if (interactive()) {
  moduleCohortOverlap(result)
}
}
