% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortTiming.R
\name{moduleCohortTiming}
\alias{moduleCohortTiming}
\title{moduleCohortTiming}
\usage{
moduleCohortTiming(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortTiming()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleCohortTiming
}
\examples{
if (interactive()) {
  moduleCohortTiming(result)
}
}
