% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DoseCoverage.R
\name{moduleDoseCoverage}
\alias{moduleDoseCoverage}
\title{moduleDoseCoverage}
\usage{
moduleDoseCoverage(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseDoseCoverage} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleDoseCoverage
}
\examples{
if (interactive()) {
  moduleDoseCoverage(result)
}
}
