% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DrugRestart.R
\name{moduleDrugRestart}
\alias{moduleDrugRestart}
\title{moduleDrugRestart}
\usage{
moduleDrugRestart(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseDrugRestart} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleDrugRestart
}
\examples{
if (interactive()) {
  moduleDrugRestart(result)
}
}
