% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DrugUtilisation.R
\name{moduleDrugUtilisation}
\alias{moduleDrugUtilisation}
\title{moduleDrugUtilisation}
\usage{
moduleDrugUtilisation(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseDrugUtilisation} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleDrugUtilisation
}
\examples{
if (interactive()) {
  moduleDrugUtilisation(result)
}
}
