% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePrevalence-IncidencePrevalence.R
\name{moduleIncidence}
\alias{moduleIncidence}
\title{moduleIncidence}
\usage{
moduleIncidence(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{estimateIncidence()} function from the \code{IncidencePrevalence} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleIncidence
}
\examples{
if (interactive()) {
  moduleIncidence(result)
}
}
