% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-Indication.R
\name{moduleIndication}
\alias{moduleIndication}
\title{moduleIndication}
\usage{
moduleIndication(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseIndication} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleIndication
}
\examples{
if (interactive()) {
  moduleIndication(result)
}
}
