% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/CohortCharacteristics-LargeScaleCharacteristics.R
\name{moduleLargeScaleCharacteristics}
\alias{moduleLargeScaleCharacteristics}
\title{moduleLargeScaleCharacteristics}
\usage{
moduleLargeScaleCharacteristics(result, .softValidation)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseLargeScaleCharacteristics()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleLargeScaleCharacteristics
}
\examples{
if (interactive()) {
  moduleLargeScaleCharacteristics(result)
}
}
