% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePrevalence-IncidencePrevalence.R
\name{modulePrevalence}
\alias{modulePrevalence}
\title{modulePrevalence}
\usage{
modulePrevalence(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from either \code{estimatePeriodPrevalence()} or \code{estimatePointPrevalence()} function from the \code{IncidencePrevalence} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
modulePrevalence
}
\examples{
if (interactive()) {
  modulePrevalence(result)
}
}
