% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSurvival-CohortSurvival.R
\name{moduleSurvival}
\alias{moduleSurvival}
\title{moduleSurvival}
\usage{
moduleSurvival(result, .softValidation)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from either \code{estimateCompetingRiskSurvival()} or \code{estimateSingleEventSurvival()} functions from the CohortSurvival package.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{TreatmentPatterns} ShinyModule
}
\description{
Wrapper function to create a CohortSurvival module instance.
}
\examples{
if (interactive()) {
  moduleSurvival(tpr)
}
}
