% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-Treatment.R
\name{moduleTreatment}
\alias{moduleTreatment}
\title{moduleTreatment}
\usage{
moduleTreatment(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseTreatment} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
moduleTreatment
}
\examples{
if (interactive()) {
  moduleTreatment(result)
}
}
