% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-TreatmentPatterns.R
\name{moduleTreatmentPatterns}
\alias{moduleTreatmentPatterns}
\title{moduleTreatmentPatterns}
\usage{
moduleTreatmentPatterns(..., .softValidation = FALSE)
}
\arguments{
\item{...}{Unnamed result objects from the \code{export()} function from the TreatmentPatterns package.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{TreatmentPatterns} ShinyModule
}
\description{
Wrapper function to create a TreatmentPatterns module instance.
}
\examples{
if (interactive()) {
  moduleTreatmentPatterns(tpr1, tpr2, tpr3)
}
}
