% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPlotly.R
\name{PlotPlotly}
\alias{PlotPlotly}
\title{PlotPlotly}
\description{
Plotly Module
}
\examples{
library(DarwinShinyModules)

plotlyFun <- function(data) {
  plotly::ggplotly(
    ggplot(data = data, mapping = aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
      geom_point() +
      theme_bw()
  )
}

plotlyModule <- PlotPlotly$new(data = iris, fun = plotlyFun)

if (interactive()) {
  preview(plotlyModule)
}
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:Plot]{DarwinShinyModules::Plot}} -> \code{PlotPlotly}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{plot}}{(\code{plotly}) object.}

\item{\code{source}}{(\code{character}) Source label for the plotly plot.}

\item{\code{bindings}}{(\code{reactivevalues}) bindings from the plotly object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotPlotly-new}{\code{PlotPlotly$new()}}
\item \href{#method-PlotPlotly-UI}{\code{PlotPlotly$UI()}}
\item \href{#method-PlotPlotly-server}{\code{PlotPlotly$server()}}
\item \href{#method-PlotPlotly-clone}{\code{PlotPlotly$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Plot" data-id="validate"><a href='../../DarwinShinyModules/html/Plot.html#method-Plot-validate'><code>DarwinShinyModules::Plot$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotPlotly-new"></a>}}
\if{latex}{\out{\hypertarget{method-PlotPlotly-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotPlotly$new(data, fun)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}

\item{\code{fun}}{Function to plot with, with one argument: \code{data}.}

\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotPlotly-UI"></a>}}
\if{latex}{\out{\hypertarget{method-PlotPlotly-UI}{}}}
\subsection{Method \code{UI()}}{
UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotPlotly$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{shiny.tag.list}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotPlotly-server"></a>}}
\if{latex}{\out{\hypertarget{method-PlotPlotly-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotPlotly$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotPlotly-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotPlotly-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotPlotly$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
