% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence.R
\name{Prevalence}
\alias{Prevalence}
\title{Prevalence}
\description{
Prevalence Module
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{Prevalence}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Prevalence-new}{\code{Prevalence$new()}}
\item \href{#method-Prevalence-UI}{\code{Prevalence$UI()}}
\item \href{#method-Prevalence-server}{\code{Prevalence$server()}}
\item \href{#method-Prevalence-clone}{\code{Prevalence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prevalence-new"></a>}}
\if{latex}{\out{\hypertarget{method-Prevalence-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prevalence$new(appId, prevalence)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prevalence-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Prevalence-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prevalence$UI(title = "Table")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prevalence-server"></a>}}
\if{latex}{\out{\hypertarget{method-Prevalence-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prevalence$server(input, output, session)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prevalence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Prevalence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prevalence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
