% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-ProportionOfPatientsCovered.R
\name{ProportionOfPatientsCovered}
\alias{ProportionOfPatientsCovered}
\title{ProportionOfPatientsCovered Module Class}
\description{
ProportionOfPatientsCovered module that shows tables and plots
}
\examples{
{
if (interacitve()) {
  library(DarwinShinyModules)
  library(DrugUtilisation)

  cdm <- DrugUtilisation::mockDrugUtilisation(numberIndividuals = 100)

  res <- cdm$cohort1 |>
    DrugUtilisation::summariseProportionOfPatientsCovered(followUpDays = 365)

  mod <- ProportionOfPatientsCovered$new(res)
  preview(mod)
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{ProportionOfPatientsCovered}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result}) Object created by \code{DrugUtilisation::summariseProportionOfPatientsCovered()}.}

    \item{\code{table}}{(\code{Flextable}) ShinyModule}

    \item{\code{plot}}{(\code{PlotStatic}) ShinyModule}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ProportionOfPatientsCovered-initialize}{\code{ProportionOfPatientsCovered$new()}}
    \item \href{#method-ProportionOfPatientsCovered-clone}{\code{ProportionOfPatientsCovered$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProportionOfPatientsCovered-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ProportionOfPatientsCovered-initialize}{}}}
\subsection{\code{ProportionOfPatientsCovered$new()}}{
  Initializer method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProportionOfPatientsCovered$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Object created by \code{DrugUtilisation::summariseProportionOfPatientsCovered()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ProportionOfPatientsCovered-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ProportionOfPatientsCovered-clone}{}}}
\subsection{\code{ProportionOfPatientsCovered$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ProportionOfPatientsCovered$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
