% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sankey.R
\name{Sankey}
\alias{Sankey}
\title{Sankey}
\description{
Sankey diagram Module.
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:TreatmentPatterns]{DarwinShinyModules::TreatmentPatterns}} -> \code{Sankey}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colours}}{(\code{list()}) Named list of hex colour codes i.e.: \code{list(A = "#FF0000", B = "#00FF00")}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sankey-clone}{\code{Sankey$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="UI"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-UI'><code>DarwinShinyModules::TreatmentPatterns$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="initialize"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-initialize'><code>DarwinShinyModules::TreatmentPatterns$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="server"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-server'><code>DarwinShinyModules::TreatmentPatterns$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="validate"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-validate'><code>DarwinShinyModules::TreatmentPatterns$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sankey-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sankey-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sankey$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
