% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveAppStructure.R
\name{saveAppStructure}
\alias{saveAppStructure}
\title{saveAppStructure}
\usage{
saveAppStructure(appStructure, filePath = "./appStructure.qs", ...)
}
\arguments{
\item{appStructure}{(\code{list}: \code{NULL}) appStructure object, i.e. a named list
of ShinyModule objects or (named) lists containing (named) ShinyModule
objects. Representing the navigation (side) bar of the shiny app.}

\item{filePath}{(\code{character}: \code{"./appStructure.qs"}) Path to the qs-file,
containing the \code{appStructure}, or where to write the \code{appStructure} to.}

\item{...}{Extra arguments for \code{qs2::qs_save()}}
}
\value{
\code{NULL} invisible
}
\description{
Wrapper for \code{qs2::qs_save()}. Saves an \code{appStructure} object as a qs-file
using \code{qs2}.
}
