% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortAttrition.R
\name{shinyCohortAttrition}
\alias{shinyCohortAttrition}
\title{shinyCohortAttrition}
\usage{
shinyCohortAttrition(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortAttrition} function from the CohortCharacteristics pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{shiny.appobj}
}
\description{
shinyCohortAttrition
}
\examples{
if (interactive()) {
  shinyCohortAttrition(result)
}
}
