% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortCodelist.R
\name{shinyCohortCodelist}
\alias{shinyCohortCodelist}
\title{shinyCohortCodelist}
\usage{
shinyCohortCodelist(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortCodelist()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{shiny.appojb}
}
\description{
shinyCohortCodelist
}
\examples{
if (interactive()) {
  moduleCohortCodelist(result)
}
}
