% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortCharacteristics-CohortCount.R
\name{shinyCohortCount}
\alias{shinyCohortCount}
\title{shinyCohortCount}
\usage{
shinyCohortCount(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseCohortCount()} function from the \code{CohortCharacteristics} pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{shiny.appojb}
}
\description{
shinyCohortCount
}
\examples{
if (interactive()) {
  moduleCohortCount(result)
}
}
