% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyModule.R
\name{ShinyModule}
\alias{ShinyModule}
\title{Module Super Class}
\description{
This class is an \code{interface} and is not meant to be directly used, but to be
inherited by other Modules.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{instanceId}}{(\code{character(1)}) Random ID of 10 capitalized letters.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) appId used for namespacing.}

\item{\code{moduleName}}{(\code{character(1)}) Name of the module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ShinyModule-new}{\code{ShinyModule$new()}}
\item \href{#method-ShinyModule-validate}{\code{ShinyModule$validate()}}
\item \href{#method-ShinyModule-UI}{\code{ShinyModule$UI()}}
\item \href{#method-ShinyModule-server}{\code{ShinyModule$server()}}
\item \href{#method-ShinyModule-id}{\code{ShinyModule$id()}}
\item \href{#method-ShinyModule-clone}{\code{ShinyModule$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-new"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$new(appId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-validate"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-validate}{}}}
\subsection{Method \code{validate()}}{
Validator method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{self})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-UI"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-UI}{}}}
\subsection{Method \code{UI()}}{
Method to include a \link[shiny]{tagList} to include the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{tagList})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-server"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-server}{}}}
\subsection{Method \code{server()}}{
Method to handle the back-end.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input}) Input from the server function.}

\item{\code{output}}{(\code{output}) Output from the server function.}

\item{\code{session}}{(\code{session}) Session from the server function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{NULL})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-id"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-id}{}}}
\subsection{Method \code{id()}}{
Create an instance unique ID for referencing objects in the input and
output environments.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$id(id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{id}}{(\code{character(1)}) ID used for \code{outputId} in output() functions
and to reference in the input environment.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ShinyModule-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ShinyModule-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ShinyModule$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
