% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DoseCoverage.R,
%   R/DrugUtilisation-ProportionOfPatientsCovered.R
\name{shinyProportionOfPatientsCovered}
\alias{shinyProportionOfPatientsCovered}
\title{shinyDoseCoverage}
\usage{
shinyProportionOfPatientsCovered(result, .softValidation = FALSE)

shinyProportionOfPatientsCovered(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseProportionOfPatientsCovered} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}

\code{ShinyModule}
}
\description{
shinyDoseCoverage

shinyProportionOfPatientsCovered
}
\examples{
if (interactive()) {
  shinyProportionOfPatientsCovered(result)
}
if (interactive()) {
  shinyProportionOfPatientsCovered(result)
}
}
