% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSurvival-CohortSurvival.R
\name{shinySurvival}
\alias{shinySurvival}
\title{shinySurvival}
\usage{
shinySurvival(result, .softValidation)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from either \code{estimateCompetingRiskSurvival()} or \code{estimateSingleEventSurvival()} functions from the CohortSurvival package.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{shiny.appobj}
}
\description{
shinySurvival
}
\examples{
if (interactive()) {
  shinySurvival(result)
}
}
