% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-Treatment.R
\name{shinyTreatment}
\alias{shinyTreatment}
\title{shinyTreatment}
\usage{
shinyTreatment(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseTreatment} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
shinyTreatment
}
\examples{
if (interactive()) {
  shinyTreatment(result)
}
}
