% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-TreatmentPatterns.R
\name{shinyTreatmentPatterns}
\alias{shinyTreatmentPatterns}
\title{shinyTreatmentPatterns}
\usage{
shinyTreatmentPatterns(..., .softValidation = FALSE)
}
\arguments{
\item{...}{Unnamed result objects from the \code{export()} function from the TreatmentPatterns package.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
shiny.appobj
}
\description{
shinyTreatmentPatterns
}
\examples{
if (interactive()) {
  shinyTreatmentPatterns(tpr1, tpr2, tpr3)
}
}
