% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\title{Table}
\description{
Table Module
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{Table}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{bindings}}{(\code{reactivevalues}) Bindings of the DataTable in a reactive environment.}

\item{\code{data}}{The data to render in the DataTable, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Table-new}{\code{Table$new()}}
\item \href{#method-Table-validate}{\code{Table$validate()}}
\item \href{#method-Table-UI}{\code{Table$UI()}}
\item \href{#method-Table-server}{\code{Table$server()}}
\item \href{#method-Table-clone}{\code{Table$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-new"></a>}}
\if{latex}{\out{\hypertarget{method-Table-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$new(appId, data, options = list(scrollX = TRUE), filter = "top")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}

\item{\code{options}}{table options, by default it shows additional items next to
the table like search box, pagination, etc. Only display the table using
list(dom = '')}

\item{\code{filter}}{filter option, it can be either "none", "bottom" or "top" (default)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Table-validate}{}}}
\subsection{Method \code{validate()}}{
validate
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Table-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$UI(title = "Table")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(\code{character(1)}) Title to use for the DataTable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag.list}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-server"></a>}}
\if{latex}{\out{\hypertarget{method-Table-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Table-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
