% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{Text}
\alias{Text}
\title{Text}
\description{
Text Module
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{Text}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{mdText}}{(\code{character(n)}) Lines read from the markdown-file supplied.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Text-new}{\code{Text$new()}}
\item \href{#method-Text-UI}{\code{Text$UI()}}
\item \href{#method-Text-clone}{\code{Text$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="server"><a href='../../DarwinShinyModules/html/Module.html#method-Module-server'><code>DarwinShinyModules::Module$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-new"></a>}}
\if{latex}{\out{\hypertarget{method-Text-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$new(appId, mdFile)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{mdFile}}{(\code{character(1)}) Path to the markdown-file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Text-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{shiny.tag.list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Text-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
