% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-Treatment.R
\name{Treatment}
\alias{Treatment}
\title{Treatment Module Class}
\value{
\code{self}
}
\description{
Restart module that shows a that supports results from the
\code{summariseTreatment()} function from the DrugUtilisation package.
}
\examples{
if (interactive()) {
  cdm <- DrugUtilisation::mockDrugUtilisation()

  result <- cdm$cohort1 |>
    PatientProfiles::addAge(
      ageGroup = list(
        `0-17` = c(0, 17),
        `>=18` = c(18, Inf)
      )
    ) |>
    PatientProfiles::addSex() |>
    DrugUtilisation::summariseTreatment(
      treatmentCohortName = "cohort2",
      window = list(c(0, 30), c(31, 365)), strata = list("age_group", "sex")
    )

  mod <- Treatment$new(result = result)

  DarwinShinyModules::preview(mod)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{Treatment}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{result}}{(\code{sumamrised_result})}

\item{\code{table}}{(\code{Flextable})}

\item{\code{plot}}{(\code{PlotStatic})}

\item{\code{cdmNames}}{(\code{character(n)})}

\item{\code{cohortNames}}{(\code{character(n)})}

\item{\code{strata}}{(\code{character(n)})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Treatment-new}{\code{Treatment$new()}}
\item \href{#method-Treatment-clone}{\code{Treatment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>DarwinShinyModules::ShinyModule$getReactiveValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Treatment-new"></a>}}
\if{latex}{\out{\hypertarget{method-Treatment-new}{}}}
\subsection{Method \code{new()}}{
Initializer method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Treatment$new(result, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{result}}{(\code{summarised_result}) Object created by \code{DrugUtilisation::summariseDrugUtilisation()}.}

\item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Treatment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Treatment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Treatment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
