% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-TreatmentPathways.R
\name{TreatmentPathways}
\alias{TreatmentPathways}
\title{TreatmentPathways}
\description{
Module that displays the Treatment Pathways from the \code{TreatmentPatterns} package.
}
\examples{
{
  if (interactive()) {
    library(DarwinShinyModules)

    tpr <- TreatmentPatterns::TreatmentPatternsResults$new(
      filePath = system.file(package = "DarwinShinyModules", "dummyData/TreatmentPatterns/3.0.0/")
    )

    treatmentPathways <- TreatmentPathways$new(
      treatmentPathways = tpr$treatment_pathways,
      cdmSourceInfo = tpr$cdm_source_info
    )

    preview(treatmentPathways)
  }
}

}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{TreatmentPathways}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{colours}}{(\code{list}) Hex colour values used in the Sunburst Plots and
Sankey Diagrams.}

\item{\code{sunburst}}{(\code{PlotWidget}) Module.}

\item{\code{sankey}}{(\code{PlotWidget}) Module.}

\item{\code{inputPanel}}{(\code{InputPanel}) Module.}

\item{\code{table}}{Table displaying the \code{treatment_pathways} csv-file.}

\item{\code{sunburstOverview}}{(\code{list}) Containing Sunburst \code{PlotWidget} modules.}

\item{\code{treatmentPathways}}{(\code{data.frame})}

\item{\code{cdmSourceInfo}}{(\code{data.frame})}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreatmentPathways-new}{\code{TreatmentPathways$new()}}
\item \href{#method-TreatmentPathways-clone}{\code{TreatmentPathways$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>DarwinShinyModules::ShinyModule$getReactiveValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPathways-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPathways-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPathways$new(treatmentPathways, cdmSourceInfo, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatmentPathways}}{(\code{data.frame}) \code{treatment_pathways} field from the \code{TreatmentPatternsResult} object.}

\item{\code{cdmSourceInfo}}{(\code{data.frame}) \code{cdm_source_info} field from the \code{TreatmentPatternsResult} object.}

\item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPathways-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPathways-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPathways$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
