% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-TreatmentPatterns.R
\name{TreatmentPatterns}
\alias{TreatmentPatterns}
\title{TreatmentPatterns Module Class}
\value{
\code{self}
}
\description{
TreatmentPatterns module that shows results from the TreatmentPatterns package.
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{TreatmentPatterns}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{analyses}}{Analyses table from TreatmentPatterns}

\item{\code{treatment_pathways}}{Analyses table from TreatmentPatterns}

\item{\code{summary_event_duration}}{Analyses table from TreatmentPatterns}

\item{\code{counts_age}}{Analyses table from TreatmentPatterns}

\item{\code{counts_sex}}{Analyses table from TreatmentPatterns}

\item{\code{counts_year}}{Analyses table from TreatmentPatterns}

\item{\code{attrition}}{Analyses table from TreatmentPatterns}

\item{\code{metadata}}{Analyses table from TreatmentPatterns}

\item{\code{arguments}}{Analyses table from TreatmentPatterns}

\item{\code{cdm_source_info}}{Analyses table from TreatmentPatterns}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreatmentPatterns-new}{\code{TreatmentPatterns$new()}}
\item \href{#method-TreatmentPatterns-clone}{\code{TreatmentPatterns$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>DarwinShinyModules::ShinyModule$getReactiveValues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Unnamed TreatmentPatternsResults objects. And Additional parameters to set fields from the \code{ShinyModule} parent.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
