% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortSurvival.R
\name{CohortSurvival}
\alias{CohortSurvival}
\title{CohortSurvival Module Class}
\description{
CohortSurvival module that shows a that supports results from the
CohortSurvival package.
}
\details{
The module consists of the following:
\describe{
\item{"PlotPlotly"}{Interactive Plotly plot, visualizing the data.}
\item{"GTTable"}{gttable visualizing the tidy data}
\item{"InputPanel"}{Input panel dealing with user input}
\item{"Table"}{basic table visualizing the raw data}
}
}
\examples{
{
\donttest{
 library(DarwinShinyModules)

 if (
   require(
     "CohortSurvival",
     character.only = TRUE,
     quietly = TRUE,
     warn.conflicts = FALSE
   )
 ) {
    library(CDMConnector)
    library(CohortSurvival)

    cdm <- CohortSurvival::mockMGUS2cdm()

    MGUS_death <- estimateSingleEventSurvival(
      cdm,
      targetCohortTable = "mgus_diagnosis",
      outcomeCohortTable = "death_cohort",
      strata = list(
        c("age_group"),
        c("sex"),
        c("age_group", "sex")
      )
    )

    cs <- CohortSurvival$new(data = MGUS_death)
    if (interactive()) {
      preview(cs)
    }
  }
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{CohortSurvival}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{SummarisedResult}) Summarised result object from \code{CohortSurvival}}

\item{\code{plot}}{(\code{Plot}) Plot module.}

\item{\code{tidyTable}}{(\code{GTTable}) GTTable module}

\item{\code{table}}{(\code{Table}) Table module}

\item{\code{inputPanel}}{(\code{InputPanel}) InputPanel module}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CohortSurvival-new}{\code{CohortSurvival$new()}}
\item \href{#method-CohortSurvival-clone}{\code{CohortSurvival$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSurvival-new"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSurvival-new}{}}}
\subsection{Method \code{new()}}{
Initializer function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSurvival$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{SummarisedResults}) Summarised result object from \code{CohortSurvival}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{invisible(self)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CohortSurvival-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CohortSurvival-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CohortSurvival$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
