% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Darwin-Utils.R
\name{darwinFooter}
\alias{darwinFooter}
\title{darwinFooter}
\usage{
darwinFooter()
}
\value{
\code{tagList}
}
\description{
Creates the DARWIN EU (c) footer.
}
\examples{
ui <- shiny::fluidPage(
  darwinHeader(),
  shiny::p("Content"),
  darwinFooter()
)

server <- function(input, output, session) {}

if (interactive()) {
  shiny::shinyApp(ui, server)
}
}
