% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DarwinShinyModules.R
\docType{package}
\name{DarwinShinyModules-package}
\alias{DarwinShinyModules}
\alias{DarwinShinyModules-package}
\title{DarwinShinyModules: Repository of Shiny Modules for Darwin Result Viewers}
\description{
Install this package to access useful shiny modules for building shiny apps to explore results using the Darwin tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://darwin-eu-dev.github.io/DarwinShinyModules/}
}

}
\author{
\strong{Maintainer}: Maarten van Kessel \email{m.l.vankessel@erasmusmc.nl} (\href{https://orcid.org/0009-0006-8832-6030}{ORCID})

Authors:
\itemize{
  \item Ger Inberg \email{g.inberg@erasmusmc.nl} (\href{https://orcid.org/0000-0001-8993-8748}{ORCID})
}

Other contributors:
\itemize{
  \item Erasmus MC [copyright holder]
}

}
\keyword{internal}
