% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatabaseDBI.R
\name{DatabaseDBI}
\alias{DatabaseDBI}
\title{DatabaseDBI Class}
\description{
This class extends the \code{Database} class, to query a database using
\code{DBI} with ODBC.
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:Database]{DarwinShinyModules::Database}} -> \code{DatabaseDBI}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{tables}}{(\code{list(n)}) List of \code{tbl_dbi} tables for the respective driver.}

\item{\code{connectArgs}}{(\code{list(n)}) Named list of additional arguments used in \code{DBI::dbConnect()}}

\item{\code{disconnectArgs}}{(\code{list(n)}) Named list of additional arguments used in \code{DBI::dbDisconnect()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DatabaseDBI-new}{\code{DatabaseDBI$new()}}
\item \href{#method-DatabaseDBI-attachTables}{\code{DatabaseDBI$attachTables()}}
\item \href{#method-DatabaseDBI-detatchTables}{\code{DatabaseDBI$detatchTables()}}
\item \href{#method-DatabaseDBI-clone}{\code{DatabaseDBI$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseDBI-new"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseDBI-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseDBI$new(driver)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{driver}}{Driver to use to connect to the database with \code{DBI::dbConnect()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{invisible(self)}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseDBI-attachTables"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseDBI-attachTables}{}}}
\subsection{Method \code{attachTables()}}{
Method to attach tables from the database in the \code{tables} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseDBI$attachTables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{character()}) Names of tables to attach.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseDBI-detatchTables"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseDBI-detatchTables}{}}}
\subsection{Method \code{detatchTables()}}{
Method to detatch tables from the \code{tables} field.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseDBI$detatchTables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{(\code{character()}) Names of the tables to detatch.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DatabaseDBI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DatabaseDBI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DatabaseDBI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
