% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-deploy.R
\name{deployAppStructure}
\alias{deployAppStructure}
\title{deployAppStructure}
\usage{
deployAppStructure(
  appStructure,
  appDir = tempfile(),
  launchFun = "launchDarwinBslibApp",
  daemons = 4,
  pkgs = NULL,
  ...
)
}
\arguments{
\item{appStructure}{(\code{list}: \code{NULL}) appStructure object, i.e. a named list
of ShinyModule objects or (named) lists containing (named) ShinyModule
objects. Representing the navigation (side) bar of the shiny app.}

\item{appDir}{(\code{character(1)}: \code{tempfile()}) File path to the directory where app files will
be written to, and the app is being deployed from. See \code{rsconnect::deployApp}
for further details.}

\item{launchFun}{(\code{character(1)}: \code{"launchDarwinBslibApp"}) Launch function as character}

\item{daemons}{(\code{numeric(1)}: \code{4}) Number of daemons (sub-processes) to launch with the shiny app.}

\item{pkgs}{(\code{character(n)}: \code{NULL}) Additional packages that should be
installed and loaded when deploying.}

\item{...}{Arguments for \code{rsconnect::deployApp}}
}
\value{
\code{NULL}
}
\description{
Wrapper for \code{rsconnect::deployApp}. Deploys a shiny app from the appStructure
}
