% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-DrugUtilisation.R
\name{DrugUtilisation}
\alias{DrugUtilisation}
\title{DrugUtilisation Module Class}
\description{
DrugUtilisation module that shows a that supports results from the
\code{summariseDrugutilisation()} function from the DrugUtilisation package.
}
\examples{
if (interactive()) {
 library(DarwinShinyModules)

 if (
   require(
     "DrugUtilisation",
     character.only = TRUE,
     quietly = TRUE,
     warn.conflicts = FALSE
   )
 ) {
    cdm <- DrugUtilisation::mockDrugUtilisation(numberIndividual = 100, source = "duckdb")

    cdm <- DrugUtilisation::generateIngredientCohortSet(
      cdm = cdm,
      name = "dus_cohort",
      ingredient = "acetaminophen",
      gapEra = 7
    )

    result <- cdm$dus_cohort |>
      PatientProfiles::addAge(ageGroup = list(
        `0-17` = c(0, 17),
        `>=18` = c(18, Inf)
      )) |>
      PatientProfiles::addSex() |>
      DrugUtilisation::summariseDrugUtilisation(
        ingredientConceptId = 1125315,
        gapEra = 7,
        strata = list("age_group", "sex")
      )

    duMod <- DrugUtilisation$new(result = result)
    preview(duMod)
  }
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{ShinyModule}} -> \code{DrugUtilisation}
}
\section{Active bindings}{
  \if{html}{\out{<div class="r6-active-bindings">}}
  \describe{
    \item{\code{result}}{(\code{summarised_result})}

    \item{\code{table}}{(\code{Flextable})}

    \item{\code{plot}}{(\code{PlotStatic})}

    \item{\code{cdmNames}}{(\code{character(n)})}

    \item{\code{cohortNames}}{(\code{character(n)})}

    \item{\code{strata}}{(\code{character(n)})}

    \item{\code{variables}}{(\code{character(n)})}

    \item{\code{estimates}}{(\code{character(n)})}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-DrugUtilisation-initialize}{\code{DrugUtilisation$new()}}
    \item \href{#method-DrugUtilisation-clone}{\code{DrugUtilisation$clone()}}
  }
}
\if{html}{\out{<details open><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>ShinyModule$UI()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="getReactiveValues"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-getReactiveValues'><code>ShinyModule$getReactiveValues()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>ShinyModule$server()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>ShinyModule$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrugUtilisation-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-DrugUtilisation-initialize}{}}}
\subsection{\code{DrugUtilisation$new()}}{
  Initializer method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DrugUtilisation$new(result, ...)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{result}}{(\code{summarised_result}) Object created by \code{DrugUtilisation::summariseDrugUtilisation()}.}
      \item{\code{...}}{Additional parameters to set fields from the \code{ShinyModule} parent.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    \code{self}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrugUtilisation-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrugUtilisation-clone}{}}}
\subsection{\code{DrugUtilisation$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{DrugUtilisation$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
