% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePrevalence.R
\name{IncidencePrevalence}
\alias{IncidencePrevalence}
\title{IncidencePrevalence Module Class}
\value{
\code{self}
}
\description{
IncidencePrevalence module that shows a that supports incidence; point and
period prevalence results from the IncidencePrevalence package.
}
\details{
The module consists of the following:
\describe{
\item{"PlotPlotly"}{Interactive Plotly plot, visualizing the data.}
\item{"GTTable"}{gttable visualizing the tidy data}
\item{"GTTable"}{gttable visualizing the attirtion data}
\item{"Table"}{basic table visualizing the raw data}
}
}
\examples{
{
\donttest{
 library(DarwinShinyModules)

 if (require("TreatmentPatterns", character.only = TRUE, quietly = TRUE, warn.conflicts = FALSE)) {
    inc <- readRDS(system.file(package = "DarwinShinyModules", "dummyData/IncidencePrevalence/rds/incidence.rds"))
    pointPrev <- readRDS(system.file(package = "DarwinShinyModules", "dummyData/IncidencePrevalence/rds/pointPrevalence.rds"))
    periodPrev <- readRDS(system.file(package = "DarwinShinyModules", "dummyData/IncidencePrevalence/rds/periodPrevalence.rds"))

    incMod <- IncidencePrevalence$new(data = inc)
    pointPrevMod <- IncidencePrevalence$new(data = pointPrev)
    periodPrevMod <- IncidencePrevalence$new(data = periodPrev)

    ui <- shiny::fluidPage(
      incMod$UI(),
      pointPrevMod$UI(),
      periodPrevMod$UI()
    )

    server <- function(input, output, session) {
      incMod$server(input, output, session)
      pointPrevMod$server(input, output, session)
      periodPrevMod$server(input, output, session)
    }

    if (interactive()) {
      shiny::shinyApp(ui = ui, server = server)
    }
  }
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{IncidencePrevalence}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{summarised_result}) Summarised result object.}

\item{\code{dataType}}{(\code{character(1)}) Assumed data type of the provided data.
One of: \code{"Incidence"}, \code{"Point Prevalence"}, or \code{"Period Prevalence"}}

\item{\code{plotPlotly}}{(\code{PlotPlotly}) Module.}

\item{\code{gtTable}}{(\code{GTTable}) Module.}

\item{\code{gtAttrition}}{(\code{GTTable}) Module.}

\item{\code{table}}{(\code{Table}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IncidencePrevalence-new}{\code{IncidencePrevalence$new()}}
\item \href{#method-IncidencePrevalence-clone}{\code{IncidencePrevalence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-new"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{(\code{summarised_result}) Result object from the \code{IncidencePrevalence} package.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
