% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePrevalence.R
\name{IncidencePrevalence}
\alias{IncidencePrevalence}
\title{IncidencePrevalence Module Class}
\description{
IncidencePrevalence super class. Composed of the \code{Plot} and \code{Table} modules.
This class is a \code{decorator} and is not meant to be directly used, but to be
inherited.
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{IncidencePrevalence}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{(\code{data.frame}) Data the \code{table} and \code{plot} fields are based on.}

\item{\code{table}}{(\code{Table}) Module.}

\item{\code{plot}}{(\code{PlotPlotly}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IncidencePrevalence-new}{\code{IncidencePrevalence$new()}}
\item \href{#method-IncidencePrevalence-validate}{\code{IncidencePrevalence$validate()}}
\item \href{#method-IncidencePrevalence-clone}{\code{IncidencePrevalence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-new"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Incidence or Prevalence data from
\code{IncidencePrevalence::estimateIncidence()} or
\code{IncidencePrevalence::estimatePrevalence()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-validate"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-validate}{}}}
\subsection{Method \code{validate()}}{
Validation method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{self})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
