% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IncidencePrevalence.R
\name{IncidencePrevalence}
\alias{IncidencePrevalence}
\title{IncidencePrevalence}
\description{
IncidencePrevalence super class. Composed of the Plot and Table modules.
This class is an \code{interface} and is not meant to be directly used, but to be
inherited.
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{IncidencePrevalence}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{table}}{(\code{Table}) Module.}

\item{\code{plot}}{(\code{PlotPlotly}) Module.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-IncidencePrevalence-new}{\code{IncidencePrevalence$new()}}
\item \href{#method-IncidencePrevalence-UI}{\code{IncidencePrevalence$UI()}}
\item \href{#method-IncidencePrevalence-server}{\code{IncidencePrevalence$server()}}
\item \href{#method-IncidencePrevalence-clone}{\code{IncidencePrevalence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-new"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-UI"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-UI}{}}}
\subsection{Method \code{UI()}}{
UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$UI(title = "Table")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(\code{character(1)}) Title to use for the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag.list}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-server"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-IncidencePrevalence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-IncidencePrevalence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{IncidencePrevalence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
