% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeModule.R
\name{makeModule}
\alias{makeModule}
\title{makeModule}
\usage{
makeModule(ui, server, namespace = NULL, data = NULL)
}
\arguments{
\item{ui}{Shiny UI elements i.e. a \code{shiny.tag.list}, or similar ui objects
from packages like \code{shiny}, \code{shinydashboard}, or \code{bslib}}

\item{server}{(\code{function}) A server function with atleast a \code{input} and
\code{output} argument. \code{data} is always availible in the server function, even
when you do not explicitly list it as an argument in the \code{server} function
definition.}

\item{namespace}{(\code{character}: \code{NULL}) Namespace used in the ui element.}

\item{data}{(\code{NULL}) Data to use. The data will be stored in the returned
module object. May be of any type that you support in your defined \code{server}
function. \code{data} is always available in your \code{sever} function definition
even if you do not explicitly pass it as an argument. If you need to use
multiple data structures in your module, you can pack them up in a \code{list}.
See the examples.}
}
\value{
\code{ShinyModule}
}
\description{
Function to make a \code{ShinyModule} from an UI element and server function.
}
\details{
The function allows for easy migration between bespoke shiny code and the
modular framework in \code{DarwinShinyModules}, without having to implement an
\code{R6} class. One caveat is, is the generated module is completely isolated.
Meaning that the module does not allow other modules to read from or write
to any defined (reactive) variables in the provided server function.
}
\examples{
library(DarwinShinyModules)
library(shiny)

ui <- fluidPage(
  tableOutput(NS("myMod", "table"))
)

# Notice that we do NOT pass a `data` argument, but it is still available in
# the `server` function
server <- function(input, output, session) {
  output$table <- renderTable({
    data
  })
}

mod <- makeModule(ui, server, namespace = "myMod", data = iris)

if (interactive()) {
  preview(mod)
}

# Multiple data structures:
ui <- fluidPage(
  tableOutput(NS("myMod", "tableIris")),
  tableOutput(NS("myMod", "tableCars"))
)

server <- function(input, output, session) {
  datIris <- data$iris
  datCars <- data$mtcars

  output$tableIris <- renderTable({
    datIris
  })

  output$tableCars <- renderTable({
    datCars
  })
}

mod <- makeModule(ui, server, namespace = "myMod", data = list(iris = iris, mtcars = mtcars))

if (interactive()) {
  preview(mod)
}
}
