% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns-TreatmentPatterns.R
\name{moduleTreatmentPatterns}
\alias{moduleTreatmentPatterns}
\title{moduleTreatmentPatterns}
\usage{
moduleTreatmentPatterns(...)
}
\arguments{
\item{...}{Unnamed TreatmentPatternsResults objects.}
}
\value{
\code{TreatmentPatterns} ShinyModule
}
\description{
Wrapper function to create a TreatmentPatterns module instance.
}
\examples{
if (interactive()) {
  moduleTreatmentPatterns(tpr)
}
}
