% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotStatic.R
\name{PlotStatic}
\alias{PlotStatic}
\title{PlotStatic}
\description{
Static Plot (\code{plot()} or \code{ggplot2}) Module
}
\examples{
library(DarwinShinyModules)
library(ggplot2)

staticFun <- function(data) {
  ggplot(data = data, mapping = aes(x = Sepal.Length, y = Sepal.Width, color = Species)) +
    geom_point() +
    theme_bw()
}

staticModule <- PlotStatic$new(data = iris, fun = staticFun)

if (interactive()) {
  preview(staticModule)
}
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:Plot]{DarwinShinyModules::Plot}} -> \code{PlotStatic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotStatic-UI}{\code{PlotStatic$UI()}}
\item \href{#method-PlotStatic-server}{\code{PlotStatic$server()}}
\item \href{#method-PlotStatic-clone}{\code{PlotStatic$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="initServer"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-initServer'><code>DarwinShinyModules::ShinyModule$initServer()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Plot" data-id="initialize"><a href='../../DarwinShinyModules/html/Plot.html#method-Plot-initialize'><code>DarwinShinyModules::Plot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Plot" data-id="validate"><a href='../../DarwinShinyModules/html/Plot.html#method-Plot-validate'><code>DarwinShinyModules::Plot$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotStatic-UI"></a>}}
\if{latex}{\out{\hypertarget{method-PlotStatic-UI}{}}}
\subsection{Method \code{UI()}}{
UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotStatic$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{shiny.tag.list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotStatic-server"></a>}}
\if{latex}{\out{\hypertarget{method-PlotStatic-server}{}}}
\subsection{Method \code{server()}}{
server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotStatic$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotStatic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotStatic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotStatic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
