% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotWidget.R
\name{PlotWidget}
\alias{PlotWidget}
\title{PlotWidget}
\description{
Widget Module
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:Plot]{DarwinShinyModules::Plot}} -> \code{PlotWidget}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PlotWidget-UI}{\code{PlotWidget$UI()}}
\item \href{#method-PlotWidget-server}{\code{PlotWidget$server()}}
\item \href{#method-PlotWidget-clone}{\code{PlotWidget$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Plot" data-id="initialize"><a href='../../DarwinShinyModules/html/Plot.html#method-Plot-initialize'><code>DarwinShinyModules::Plot$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Plot" data-id="updateDataReactive"><a href='../../DarwinShinyModules/html/Plot.html#method-Plot-updateDataReactive'><code>DarwinShinyModules::Plot$updateDataReactive()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotWidget-UI"></a>}}
\if{latex}{\out{\hypertarget{method-PlotWidget-UI}{}}}
\subsection{Method \code{UI()}}{
UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotWidget$UI(title = "Widget")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(\code{character(1)}) Title to use for the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag.list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotWidget-server"></a>}}
\if{latex}{\out{\hypertarget{method-PlotWidget-server}{}}}
\subsection{Method \code{server()}}{
server
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotWidget$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PlotWidget-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PlotWidget-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PlotWidget$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
