% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prevalence.R
\name{Prevalence}
\alias{Prevalence}
\title{Prevalence}
\description{
Prevalence Module
}
\examples{
library(DarwinShinyModules)

ipInstalled <- require(
  "IncidencePrevalence",
  character.only = TRUE,
  quietly = TRUE,
  warn.conflicts = FALSE
)

if (ipInstalled) {
  prev <- readRDS(system.file(
    package = "DarwinShinyModules",
    "dummyData/IncidencePrevalence/rds/prevalence.rds"
  ))

  prevalence <- Prevalence$new(data = prev)
}

if (interactive()) {
  preview(prevalence)
}
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:IncidencePrevalence]{DarwinShinyModules::IncidencePrevalence}} -> \code{Prevalence}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Prevalence-clone}{\code{Prevalence$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="IncidencePrevalence" data-id="UI"><a href='../../DarwinShinyModules/html/IncidencePrevalence.html#method-IncidencePrevalence-UI'><code>DarwinShinyModules::IncidencePrevalence$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="IncidencePrevalence" data-id="initialize"><a href='../../DarwinShinyModules/html/IncidencePrevalence.html#method-IncidencePrevalence-initialize'><code>DarwinShinyModules::IncidencePrevalence$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="IncidencePrevalence" data-id="server"><a href='../../DarwinShinyModules/html/IncidencePrevalence.html#method-IncidencePrevalence-server'><code>DarwinShinyModules::IncidencePrevalence$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="IncidencePrevalence" data-id="validate"><a href='../../DarwinShinyModules/html/IncidencePrevalence.html#method-IncidencePrevalence-validate'><code>DarwinShinyModules::IncidencePrevalence$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Prevalence-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Prevalence-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Prevalence$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
