% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview}
\usage{
preview(module)
}
\arguments{
\item{module}{(\link{ShinyModule}) A module object.}
}
\value{
\code{NULL}
}
\description{
Launches a shiny app with the modules' \code{server()} and \code{UI()} methods.
}
\examples{
table <- Table$new(appId = "id", data = iris)

if (interactive()) {
  preview(table)
}
}
