% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-ProportionOfPatientsCovered.R
\name{shinyProportionOfPatientsCovered}
\alias{shinyProportionOfPatientsCovered}
\title{shinyProportionOfPatientsCovered}
\usage{
shinyProportionOfPatientsCovered(result, .softValidation = FALSE)
}
\arguments{
\item{result}{(\code{summarised_result}) Result from the \code{summariseDrugRestart} function from the DrugUtilisation pacakge.}

\item{.softValidation}{(\code{logical(1)}: \code{FALSE}) When \code{TRUE} will throw the failed check as a warning.}
}
\value{
\code{ShinyModule}
}
\description{
shinyProportionOfPatientsCovered
}
\examples{
if (interactive()) {
  shinyDrugRestart(result)
}
}
