% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sunburst.R
\name{Sunburst}
\alias{Sunburst}
\title{Sunburst}
\description{
Sunburst Module.
}
\section{Super classes}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{\link[DarwinShinyModules:TreatmentPatterns]{DarwinShinyModules::TreatmentPatterns}} -> \code{Sunburst}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{colours}}{(\code{list()}) Named list of names (domain) and hex colour
codes (range):\cr \code{list(domain = c("A", "B"), range = c("#FF0000", "#00FF00"))}.
See \link[sunburstR]{sunburst}}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{jsShowLegend}}{(\code{character(1)}) JavaScript function to show legend
on render as text.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sunburst-clone}{\code{Sunburst$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="UI"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-UI'><code>DarwinShinyModules::TreatmentPatterns$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="initialize"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-initialize'><code>DarwinShinyModules::TreatmentPatterns$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="server"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-server'><code>DarwinShinyModules::TreatmentPatterns$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="TreatmentPatterns" data-id="validate"><a href='../../DarwinShinyModules/html/TreatmentPatterns.html#method-TreatmentPatterns-validate'><code>DarwinShinyModules::TreatmentPatterns$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sunburst-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Sunburst-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sunburst$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
