% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Table.R
\name{Table}
\alias{Table}
\title{Table}
\description{
Table Module
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{Table}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{dataTableBindings}}{(\code{reactivevalues}) Bindings of the DataTable in a reactive environment.}

\item{\code{data}}{The data to render in the DataTable, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Table-new}{\code{Table$new()}}
\item \href{#method-Table-validate}{\code{Table$validate()}}
\item \href{#method-Table-UI}{\code{Table$UI()}}
\item \href{#method-Table-server}{\code{Table$server()}}
\item \href{#method-Table-clone}{\code{Table$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-new"></a>}}
\if{latex}{\out{\hypertarget{method-Table-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}

\item{\code{fun}}{Function to plot with, with one argument: \code{data}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
validate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Table-validate}{}}}
\subsection{Method \code{validate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{self}
UI
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Table-UI}{}}}
\subsection{Method \code{UI()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$UI(title = "Table")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{title}}{(\code{character(1)}) Title to use for the DataTable.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{shiny.tag.list}
server
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-server"></a>}}
\if{latex}{\out{\hypertarget{method-Table-server}{}}}
\subsection{Method \code{server()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})}

\item{\code{output}}{(\code{output})}

\item{\code{session}}{(\code{session})}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Table-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Table-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Table$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
