% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Text.R
\name{Text}
\alias{Text}
\title{Text}
\description{
Text Module
}
\examples{
library(DarwinShinyModules)

md <- c(
  "# H1
  ## H2
  ### H3",

  "**bold text**",

  "*italicized text*",

  "> blockquote",

  "1. First item
  2. Second item
  3. Third item",

  "- First item
  - Second item
  - Third item",

  "`code`",

  "---",

  "[link](https://www.markdownguide.org/cheat-sheet/)",

  "![alt text](https://mdg.imgix.net/assets/images/
   san-juan-mountains.jpg?auto=format&fit=clip&q=40&w=1080)",

  "| Syntax | Description |
  | ----------- | ----------- |
  | Header | Title |
  | Paragraph | Text |",

  "```r
    foo <- function(bar, baz) {
      return(bar ** baz)
    }

    foo(2, 3)
  ```",

  "	Here's a sentence with a footnote. [^1]

  [^1]: This is the footnote.",

  "### My Great Heading {#custom-id}",

  "term
  : definition",

  "~~The world is flat.~~",

  "- [x] Write the press release
  - [ ] Update the website
  - [ ] Contact the media",

  "That is so funny! :joy:",

  "I need to highlight these ==very important words==.",
  "H~2~O",
  "X^2^"
)

text <- Text$new(appId = "app", markdown = md)

if (interactive()) {
  preview(module = text)
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{Text}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{markdown}}{(\code{character(n)}) Lines read from the markdown-file supplied.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Text-new}{\code{Text$new()}}
\item \href{#method-Text-validate}{\code{Text$validate()}}
\item \href{#method-Text-UI}{\code{Text$UI()}}
\item \href{#method-Text-clone}{\code{Text$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="id"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-id'><code>DarwinShinyModules::ShinyModule$id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-new"></a>}}
\if{latex}{\out{\hypertarget{method-Text-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$new(appId, markdown)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)}) ID of the app, to use for namespacing.}

\item{\code{markdown}}{(\code{character(n)}) Markdown.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{self}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-validate"></a>}}
\if{latex}{\out{\hypertarget{method-Text-validate}{}}}
\subsection{Method \code{validate()}}{
Validator method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{self})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-UI"></a>}}
\if{latex}{\out{\hypertarget{method-Text-UI}{}}}
\subsection{Method \code{UI()}}{
UI
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{shiny.tag.list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Text-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Text-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Text$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
