% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns.R
\name{TreatmentPatterns}
\alias{TreatmentPatterns}
\title{TreatmentPatterns Module Class}
\description{
TreatmentPatterns super class. Composed of a \code{PlotWidget} and \code{Table} modules.
This class is a \code{decorator} and is not meant to be directly used, but to be
inherited.
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{TreatmentPatterns}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Underlying data the \code{table} and \code{widget} fields are based on.}

\item{\code{inputPanel}}{(\link{InputPanel}) Module}

\item{\code{widget}}{(\link{PlotWidget}) Module}

\item{\code{table}}{(\link{Table}) Module}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreatmentPatterns-new}{\code{TreatmentPatterns$new()}}
\item \href{#method-TreatmentPatterns-validate}{\code{TreatmentPatterns$validate()}}
\item \href{#method-TreatmentPatterns-clone}{\code{TreatmentPatterns$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$new(data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{The contents of the \code{treatmentPathways.csv} file from
\code{TreatmentPatterns::export()}, as a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-validate"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-validate}{}}}
\subsection{Method \code{validate()}}{
Validation method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{self})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
