% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns.R
\name{TreatmentPatterns}
\alias{TreatmentPatterns}
\title{TreatmentPatterns Module Class}
\description{
TreatmentPatterns module that shows a Sunburst plot and Sankey diagram, with
a table.
}
\details{
The module consists of the following:
\describe{
\item{"InputPanel"}{Input panel to filter data.}
\item{"PlotWidget"}{Sunburst Plot, visualizing the data.}
\item{"PlotWidget"}{Sankey Diagram, visualizing the data.}
\item{"Table"}{Table containing the data.}
}
}
\examples{
{
library(DarwinShinyModules)

tp <- data.frame(
  path = c("A+B-C", "B+C-A", "B-A+C"),
  freq = 100,
  sex = "all",
  age = "all",
  indexYear = "all"
)

treatmentPathways <- TreatmentPatterns$new(treatmentPathways = tp)

if (interactive()) {
  preview(treatmentPathways)
}
}
}
\section{Super class}{
\code{\link[DarwinShinyModules:ShinyModule]{DarwinShinyModules::ShinyModule}} -> \code{TreatmentPatterns}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{sankeyCols}}{(\code{list(a = "#ff33cc")}) Colours for the Sankey diagram.}

\item{\code{sunburstCols}}{(\code{list(domain = list(), range = list())}) Colours for the Sunburst plot.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreatmentPatterns-new}{\code{TreatmentPatterns$new()}}
\item \href{#method-TreatmentPatterns-clone}{\code{TreatmentPatterns$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="UI"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-UI'><code>DarwinShinyModules::ShinyModule$UI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="server"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-server'><code>DarwinShinyModules::ShinyModule$server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="ShinyModule" data-id="validate"><a href='../../DarwinShinyModules/html/ShinyModule.html#method-ShinyModule-validate'><code>DarwinShinyModules::ShinyModule$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$new(treatmentPathways)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{treatmentPathways}}{(\code{data.frame}) Contents of the treatmentPathways.csv file from the \code{export()} function of TreatmentPatterns.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
