% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreatmentPatterns.R
\name{TreatmentPatterns}
\alias{TreatmentPatterns}
\title{TreatmentPatterns}
\description{
TreatmentPatterns super class. Composed of a PlotWidget and Table modules.
This class is an \code{interface} and is not meant to be directly used, but to be
inherited.
}
\section{Super class}{
\code{\link[DarwinShinyModules:Module]{DarwinShinyModules::Module}} -> \code{TreatmentPatterns}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}

\item{\code{widget}}{(\link[DarwinShinyModules]{PlotWidget}) Module}

\item{\code{table}}{(\link[DarwinShinyModules]{Table} Module}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TreatmentPatterns-new}{\code{TreatmentPatterns$new()}}
\item \href{#method-TreatmentPatterns-UI}{\code{TreatmentPatterns$UI()}}
\item \href{#method-TreatmentPatterns-server}{\code{TreatmentPatterns$server()}}
\item \href{#method-TreatmentPatterns-clone}{\code{TreatmentPatterns$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="DarwinShinyModules" data-topic="Module" data-id="validate"><a href='../../DarwinShinyModules/html/Module.html#method-Module-validate'><code>DarwinShinyModules::Module$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-new"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-new}{}}}
\subsection{Method \code{new()}}{
Initializer method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$new(appId, data)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{appId}}{(\code{character(1)})}

\item{\code{data}}{Data to plot with, usually a \code{data.frame}-like object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{invisible(self)})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-UI"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-UI}{}}}
\subsection{Method \code{UI()}}{
Method to include a \link[shiny]{tagList} to include the body.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$UI()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{tagList})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-server"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-server}{}}}
\subsection{Method \code{server()}}{
Method to handle the back-end.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$server(input, output, session)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{(\code{input})\cr
Input from the server function.}

\item{\code{output}}{(\code{output})\cr
Output from the server function.}

\item{\code{session}}{(\code{session})\cr
Session from the server function.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{NULL}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TreatmentPatterns-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TreatmentPatterns-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TreatmentPatterns$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
